Replace stringy eval with can_load()
Although this is more verbose than the stringy eval equivalent code,
it doesn't use an `eval` to do its work and we get to avoid issues with
using `$@` (because it can't always be relied upon to tell if an `eval`
failed).  This way the code's a bit nicer and hopefully more robust.