Reinstate stringy eval
I found that the solution with `can_load()` only works if the `autoload`
option is set to true.  This option is only available from Perl version
5.20.  I thus decided that backwards compatibility was more important to
me than strict standards compliance, hence I've reinstated the previous
stringy eval code.

The situation seems much more complicated when passing a version number
to a module when loading it than when loading a module without version
number restrictions.  In the latter case it's possible to use the block
form of `eval` thus avoiding the whole stringy eval issue.  The
`can_load()` option would have been nice if it worked for older Perl
versions, but that didn't seem possible.